refactor(web): separate Tavily tool boundaries
